Move `ErrorPlaceholder` to `:model`
