+++ java/src/main/kotlin/io/spine/validation/java/generate/option/ChoiceGenerator.kt
@@ -28,7 +28,6 @@ package io.spine.validation.java.generate.option
 
 import io.spine.base.FieldPath
 import io.spine.protodata.ast.TypeName
-import io.spine.protodata.ast.qualifiedName
 import io.spine.protodata.java.CodeBlock
 import io.spine.protodata.java.Expression
 import io.spine.protodata.java.ReadVar
@@ -47,9 +46,9 @@ import io.spine.validation.java.generate.OptionGenerator
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentName
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.violations
-import io.spine.validation.java.violation.ErrorPlaceholder
-import io.spine.validation.java.violation.ErrorPlaceholder.GROUP_PATH
-import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.ErrorPlaceholder
+import io.spine.validation.ErrorPlaceholder.GROUP_PATH
+import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validation.java.violation.constraintViolation
 import io.spine.validation.java.violation.templateString
