+++ java/src/main/kotlin/io/spine/validation/java/generate/option/PatternGenerator.kt
@@ -66,13 +66,13 @@ import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.pa
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.violations
 import io.spine.validation.java.generate.mangled
-import io.spine.validation.java.violation.ErrorPlaceholder
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_MODIFIERS
-import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_PATTERN
+import io.spine.validation.ErrorPlaceholder
+import io.spine.validation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.ErrorPlaceholder.REGEX_MODIFIERS
+import io.spine.validation.ErrorPlaceholder.REGEX_PATTERN
 import io.spine.validation.java.violation.constraintViolation
 import io.spine.validation.java.violation.templateString
 import java.util.regex.Matcher
