+++ java/src/main/kotlin/io/spine/validation/java/generate/option/WhenGenerator.kt
@@ -58,12 +58,12 @@ import io.spine.validation.java.generate.ValidationCodeInjector.MessageScope.mes
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentName
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.generate.ValidationCodeInjector.ValidateScope.violations
-import io.spine.validation.java.violation.ErrorPlaceholder
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.WHEN_IN
+import io.spine.validation.ErrorPlaceholder
+import io.spine.validation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.ErrorPlaceholder.WHEN_IN
 import io.spine.validation.java.violation.constraintViolation
 import io.spine.validation.java.violation.templateString
