+++ java/src/main/kotlin/io/spine/validation/java/generate/option/bound/MinGenerator.kt
@@ -44,13 +44,13 @@ import io.spine.validation.java.expression.joinToString
 import io.spine.validation.java.expression.stringValueOf
 import io.spine.validation.java.generate.SingleOptionCode
 import io.spine.validation.java.generate.OptionGenerator
-import io.spine.validation.java.violation.ErrorPlaceholder
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.MIN_OPERATOR
-import io.spine.validation.java.violation.ErrorPlaceholder.MIN_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.ErrorPlaceholder
+import io.spine.validation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.ErrorPlaceholder.MIN_OPERATOR
+import io.spine.validation.ErrorPlaceholder.MIN_VALUE
+import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
 
 /**
  * The generator for `(min)` option.
