+++ java/src/main/kotlin/io/spine/validation/java/setonce/SetOnceConstraintViolation.kt
@@ -36,11 +36,11 @@ import io.spine.protodata.java.StringLiteral
 import io.spine.protodata.java.newBuilder
 import io.spine.validate.ConstraintViolation
 import io.spine.validation.IF_SET_AGAIN
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PROPOSED_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.ErrorPlaceholder.FIELD_PROPOSED_VALUE
+import io.spine.validation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validation.java.violation.constraintViolation
 import io.spine.validation.java.violation.templateString
